refactor bill view client logic
